chore: change file structure
