+++ src/components/filterPatterns/modal/props.ts
@@ -1,0 +1,9 @@
+import { FilterPatternProps } from '../props';
+import { ActionButtonProps } from '../ActionButton';
+
+export type ModalFilterProps = FilterPatternProps &
+  Omit<ActionButtonProps, 'onClose'> & {
+    onModalOpen?: () => void;
+    onModalClose?: () => void;
+    initialModalState?: 'open' | 'closed';
+  };
