+++ src/components/filterPatterns/popover/index.stories.mdx
@@ -1,8 +1,9 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
 import { action } from '@storybook/addon-actions';
 
-import CheckboxFilter from '../checkboxFilter';
-import Box from '../box';
+import CheckboxFilter from 'src/components/checkboxFilter';
+import Box from 'src/components/box';
+
 import { PopoverFilter } from './index.tsx';
 
 <Meta
