+++ src/components/filterPatterns/popover/props.ts
@@ -1,0 +1,10 @@
+import { FilterPatternProps } from '../props';
+import { ActionButtonProps } from '../ActionButton';
+
+export type PopoverFilterProps = FilterPatternProps &
+  Omit<ActionButtonProps, 'onClose'> & {
+    initialPopoverState?: 'open' | 'closed';
+
+    onPopoverClose?: () => void;
+    onPopoverOpen?: () => void;
+  };
