+++ src/components/index.ts
@@ -40,7 +40,9 @@ export { default as ListItem } from './list/ListItem';
 export { default as Menu } from './menu';
 export { default as MissingImage } from './missingImage';
 export { default as Pagination } from './pagination';
-export { PopoverFilter } from './popoverFilter';
+export { PopoverFilter } from './filterPatterns/popover';
+export { ModalFilter } from './filterPatterns/modal';
+export { default as FilterHeading } from './filterPatterns/heading';
 export { default as Radio } from './radio';
 export { default as RadioList } from './radio/RadioList';
 export { default as RangeFilterInput } from './rangeFilterInput';
